requirements.txt
matplotlib 
numpy
opencv-python
Pillow
PyYAML
requests
scipy
tqdm
seaborn
pandas
thop
shapely